Cleaned up the basque fiscalization library.
