+++ src/Basque/Mews.Fiscalizations.Basque/Model/Receiver/Receiver.cs
@@ -22,7 +22,7 @@ namespace Mews.Fiscalizations.Basque.Model
             return new Receiver(new LocalReceiver(nif, name, postalCode, address));
         }
 
-        public static Receiver Foreign(IdType idType, String1To20 id, Name name, PostalCode postalCode, String1To250 address, Country country = null)
+        public static Receiver Foreign(IdType idType, String1To20 id, Name name, PostalCode postalCode, String1To250 address, Country country)
         {
             return new Receiver(new ForeignReceiver(idType, id, name, postalCode, address, country));
         }
