+++ src/Basque/Mews.Fiscalizations.Basque/Model/Software/LocalSoftwareDeveloper.cs
@@ -2,9 +2,9 @@
 
 namespace Mews.Fiscalizations.Basque.Model
 {
-    public sealed class LocalDeveloper
+    public sealed class LocalSoftwareDeveloper
     {
-        public LocalDeveloper(TaxpayerIdentificationNumber nif)
+        public LocalSoftwareDeveloper(TaxpayerIdentificationNumber nif)
         {
             Nif = Check.IsNotNull(nif, nameof(nif));
         }
